Completed initial Item/ItemImages resources
Still have some work to do on ItemImages. Need a good way to select the
primary image (other than just using the first image) for a given item.
Also decided to get rid of the admin namespace as it was just adding an
unneeded level of complexity - simpler it seems to just render a
different form of the view for admin purposes.