Completed Primary Image Refactoring
Now all items have two types of attachments. Each item has a primary
image which is attached directly to the Item object to allow for easier
access to each item's primary image and a somewhat cleaner new/edit ux.
It also simplifies the use of primary images in the views as we don't
have to check for the presence of an image before displaying it.